Pulled some form-submission code out into the 'formsub' module
