+++ core/src/main/resources/lib/layout/layout.jelly
@@ -101,7 +101,6 @@ ${h.initPageVariables(context)}
 
     <script src="${resURL}/scripts/prototype.js" type="text/javascript"/>
     <script src="${resURL}/scripts/behavior.js" type="text/javascript"/>
-    <script src="${resURL}/bundles/jenkins.js" type="text/javascript"/>
 
     <!-- we include our own prototype.js, so don't let stapler pull in another. -->
     <st:adjunct assumes="org.kohsuke.stapler.framework.prototype.prototype"
@@ -127,6 +126,7 @@ ${h.initPageVariables(context)}
     <l:yui module="storage" />
     <!--l:yui module="editor" suffix="-beta" /-->
 
+    <script src="${resURL}/bundles/jenkins.js" type="text/javascript"/>
     <script src="${resURL}/scripts/hudson-behavior.js" type="text/javascript"></script>
     <script src="${resURL}/scripts/sortable.js" type="text/javascript"/>
